% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_bibtex.R
\name{citation_text}
\alias{citation_text}
\title{Format a manifest record's citation as prose}
\usage{
citation_text(record)
}
\arguments{
\item{record}{A manifest record.}
}
\value{
A character vector of formatted citation text.
}
\description{
Convenience wrapper over \code{\link[=as_bibentry]{as_bibentry()}} + \code{format(..., style = "text")}.
}
