% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_once}
\alias{download_once}
\title{Download a file, skipping if file already exists locally}
\usage{
download_once(url, destfile, ...)
}
\arguments{
\item{url}{character, specifying a URL to a file}

\item{destfile}{character, specifying a local destination file path}

\item{...}{additional parameters passed to \code{download.file}}
}
\description{
Download a file, skipping if file already exists locally
}
\seealso{
\code{\link[=drive_download_once]{drive_download_once()}}
}
