% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{drive_download_folder}
\alias{drive_download_folder}
\title{Download multiple files/folders from Google Drive}
\usage{
drive_download_folder(drive_folder, path, batch_size = 10, overwrite = TRUE)
}
\arguments{
\item{drive_folder}{The folder on Google Drive, given as a URL, file id, or dribble.}

\item{path}{The local destination folder.}

\item{batch_size}{the maximum number of files per upload batch.
Downloads are done in batches to mitigate curl handle errors.}

\item{overwrite}{logical indicating whether to overwrite local files}
}
\value{
Invisibly, a dribble of the downloaded files (not directories).
}
\description{
Based on \url{https://github.com/tidyverse/googledrive/issues/123#issuecomment-563484927}.
}
\details{
Recursively download the contents of a folder from Google Drive.
}
