% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{drive_download_once}
\alias{drive_download_once}
\title{Download a file from Google Drive, skipping if file already exists locally}
\usage{
drive_download_once(file, path, ...)
}
\arguments{
\item{file}{Something that identifies the file of interest on your Google
Drive. Can be a name or path, a file id or URL marked with \code{\link[googledrive:as_id]{as_id()}}, or a
\code{\link[googledrive]{dribble}}.}

\item{path}{Character. Path for output file. If absent, the default file name
is the file's name on Google Drive and the default location is working
directory, possibly with an added file extension.}

\item{...}{additional parameters passed to \link[utils:download.file]{utils::download.file}}
}
\description{
Download a file from Google Drive, skipping if file already exists locally
}
\seealso{
\code{\link[=download_once]{download_once()}}
}
