% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{drive_upload_folder}
\alias{drive_upload_folder}
\title{Upload a folder to Google Drive}
\usage{
drive_upload_folder(folder, drive_path, batch_size = 10)
}
\arguments{
\item{folder}{The local folder that is to be uploaded, given as a path e.g. with \code{fs::path()}.
The folder and its contents are uploaded.}

\item{drive_path}{The destination folder on Google Drive, given as a URL, file id, or dribble}

\item{batch_size}{the maximum number of files per upload batch.
Uploads are done in batches to mitigate curl handle errors.}
}
\value{
Invisibly, a dribble of the uploaded files (not directories though).
}
\description{
Based on \url{https://github.com/tidyverse/googledrive/issues/200#issuecomment-1112766367}.
}
\details{
Upload the contents of a folder (directory) to Google Drive recursively.
The implementation is depth-first.
Only uploads objects that have type "file" or "directory" according to \code{fs::dir_info()};
ignores types "symlink", "FIFO", "socket", "character_device" or "block_device".
}
