% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{find_modules}
\alias{find_modules}
\title{Find modules}
\usage{
find_modules(path)
}
\arguments{
\item{path}{character, specifying the path to the modules directory.}
}
\value{
character vector
}
\description{
Determine the relative path of modules discovered in \code{path}.
}
