% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_module_packages}
\alias{get_module_packages}
\title{Get module package dependencies}
\usage{
get_module_packages(module = NULL, path, verbose = FALSE)
}
\arguments{
\item{module}{character, specifying a module name.
if not specified, all modules in \code{path} will be scanned for package dependencies.}

\item{path}{character, specifying the path to the modules directory.}

\item{verbose}{logical indicating whether additional debugging messages should be printed.}
}
\value{
a \code{data.table} with package name, repo, and minimum versions
}
\description{
Get module package dependencies
}
\seealso{
\code{\link[=check_project_packages]{check_project_packages()}}
}
