% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages_from_snapshot}
\alias{packages_from_snapshot}
\title{Extract package info from a snapshot}
\usage{
packages_from_snapshot(snapshot = "renv.lock")
}
\arguments{
\item{snapshot}{character. path to a \pkg{renv} or \pkg{Require} snapshot file.}
}
\value{
a \code{data.table} object
}
\description{
Extract package info from a snapshot
}
