% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.workflow-package.R
\docType{package}
\name{project.workflow-package}
\alias{project.workflow}
\alias{project.workflow-package}
\title{project.workflow: Project Workflow Management Tools and Utilities}
\description{
Various helpers for working with large projects.
}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

}
\keyword{internal}
