% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibility_receipt.R
\name{reproducibility_receipt}
\alias{reproducibility_receipt}
\title{Reproducibility receipt for Rmarkdown documents}
\usage{
reproducibility_receipt(
  prjDir = NULL,
  title = "Reproducibility receipt",
  writeTo = NULL
)
}
\arguments{
\item{prjDir}{path to project directory}

\item{title}{Header title for the inserted details section.}

\item{writeTo}{If provided, an markdown filename to write to (e.g., \verb{outputs/<runName>/INFO.md}).
File path is assumed to be relative to \code{prjDir}.}
}
\description{
Insert git repository and R session info into Rmarkdown documents.
Based on suggestions in a Twitter thread by Miles McBain
(\url{https://twitter.com/MilesMcBain/status/1263272935197782016?s=20}).
}
\details{
Add the following to your Rmd files (without the backslashes):

\verb{
-----

\```{r details, echo=FALSE}
workflowtools::reproducibility_receipt()
\```
}
}
\seealso{
\code{\link[=info_project]{info_project()}}
}
