% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_machine}
\alias{setup_machine}
\alias{setup_project}
\title{Create a new project}
\usage{
setup_machine()

setup_project(path = NULL, repo = NULL)
}
\arguments{
\item{path}{character. path to the new project directory}

\item{repo}{character. GitHub repository specification of the form \verb{<user|org>/<repo>@<branch>}.}
}
\value{
\code{NULL}, invisibly. Used for the side-effect of printing project setup messages to screen.
}
\description{
Every project will be a little different, and will require some level of customization.
Instead of trying to do everything for user, we provide suggestions for project setup.
Most of these follow the project setup suggestions in \pkg{usethis}.
}
\details{
\itemize{
\item \code{setup_machine()} prints a set of commands recommended for first time computer setup;
\item \code{setup_project()} prints a set of commands recommended for initializing a new project.
}
}
