% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowtools-package.R
\docType{package}
\name{workflowtools-package}
\alias{workflowtools}
\alias{workflowtools-package}
\title{workflowtools: Project Workflow Management Tools and Utilities}
\description{
Various helpers for working with large projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FOR-CAST/workflowtools}
  \item Report bugs at \url{https://github.com/FOR-CAST/workflowtools/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

}
\keyword{internal}
