% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.github_repo}
\alias{.github_repo}
\alias{.github_user}
\alias{.github_ref}
\title{Extract components of a GitHub repository string}
\usage{
.github_repo(x)

.github_user(x)

.github_ref(x)
}
\arguments{
\item{x}{}
}
\value{
character string corresponding to the extracted component.
}
\description{
Extract components of a GitHub repository string
}
\keyword{internal}
