% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{project_path}
\alias{project_path}
\alias{project_name}
\title{Project root directory}
\usage{
project_path()

project_name()
}
\value{
\code{project_path()} returns an absolute path;
\code{project_name()} returns the basename of the path.
}
\description{
Searches from current working directory for an RStudio project file
or git repository, falling back on using the current working directory.
}
