% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{use_module}
\alias{use_module}
\title{Use a SpaDES module with a project}
\usage{
use_module(path = NULL, repo = NULL)
}
\arguments{
\item{path}{character specifying the module path (e.g. \code{"modules"})}

\item{repo}{character. GitHub repository specification of the form \verb{<user|org>/<repo>@<branch>}.}
}
\value{
\code{NULL} invisibly. invoked for side-effect of adding a git submodule
}
\description{
Adds
}
